+++ pkg/zeek/escape.go
@@ -23,17 +23,6 @@ func Escape(data []byte) string {
 	return string(buf)
 }
 
-func ParseEscape(data []byte) (byte, int) {
-	if len(data) >= 4 && data[1] == 'x' {
-		v1 := unhex(data[2])
-		v2 := unhex(data[3])
-		if v1 <= 0xf || v2 <= 0xf {
-			return v1<<4 | v2, 4
-		}
-	}
-	return data[1], 2
-}
-
 // Unescape is the inverse of Escape.
 func Unescape(data []byte) []byte {
 	if bytes.IndexByte(data, '\\') < 0 {
@@ -53,6 +42,17 @@ func Unescape(data []byte) []byte {
 		buf = append(buf, c)
 	}
 	return buf
+}
+
+func ParseEscape(data []byte) (byte, int) {
+	if len(data) >= 4 && data[1] == 'x' {
+		v1 := unhex(data[2])
+		v2 := unhex(data[3])
+		if v1 <= 0xf || v2 <= 0xf {
+			return v1<<4 | v2, 4
+		}
+	}
+	return data[1], 2
 }
 
 func unhex(b byte) byte {
