+++ pkg/zval/zval.go
@@ -15,132 +15,6 @@ import (
 	"fmt"
 )
 
-const (
-	beginContainer = -1
-	endContainer   = -2
-)
-
-type node struct {
-	innerLen int
-	outerLen int
-	dfs      int
-}
-
-// Builder implements an API for holding an intermediate representation
-// of a hierarchical set of values arranged in a tree, e.g., structured
-// values that can contain nested and recursive aggregate values.
-// We encode a DFS traversal in a flat data structure that can be
-// reused across invocations so we don't otherwise allocate a tree
-// data structure for every record parsed that would then be GC'd.
-type Builder struct {
-	nodes  []node
-	leaves [][]byte
-}
-
-func NewBuilder() *Builder {
-	return &Builder{
-		nodes:  make([]node, 0, 64),
-		leaves: make([][]byte, 0, 64),
-	}
-}
-
-func (b *Builder) Reset() {
-	b.nodes = b.nodes[:0]
-	b.leaves = b.leaves[:0]
-}
-
-func (b *Builder) Begin() {
-	b.nodes = append(b.nodes, node{dfs: beginContainer})
-}
-
-func (b *Builder) End() {
-	b.nodes = append(b.nodes, node{dfs: endContainer})
-}
-
-func (b *Builder) Append(leaf []byte) {
-	k := len(b.leaves)
-	b.leaves = append(b.leaves, leaf)
-	b.nodes = append(b.nodes, node{dfs: k})
-}
-
-func (b *Builder) measure(off int) int {
-	node := &b.nodes[off]
-	dfs := node.dfs
-	if dfs == beginContainer {
-		// skip over start token
-		off++
-		for off < len(b.nodes) {
-			next := b.measure(off)
-			if next < 0 {
-				// skip over end token
-				off++
-				break
-			}
-			node.innerLen += b.nodes[off].outerLen
-			off = next
-		}
-		node.outerLen = sizeOfContainer(node.innerLen)
-		return off
-	}
-	if dfs == endContainer {
-		return -1
-	}
-	n := len(b.leaves[dfs])
-	node.innerLen = n
-	node.outerLen = sizeOfValue(n)
-	return off + 1
-}
-
-func (b *Builder) encode(dst []byte, off int) ([]byte, int) {
-	node := &b.nodes[off]
-	dfs := node.dfs
-	if dfs == beginContainer {
-		// skip over start token
-		off++
-		if b.nodes[off].dfs == endContainer {
-			return AppendUvarint(dst, containerTagUnset), off + 1
-		}
-		dst = AppendUvarint(dst, containerTag(node.innerLen))
-		for off < len(b.nodes) {
-			var next int
-			dst, next = b.encode(dst, off)
-			if next < 0 {
-				// skip over end token
-				off++
-				break
-			}
-			off = next
-		}
-		return dst, off
-	}
-	if dfs == endContainer {
-		return dst, -1
-	}
-	return AppendValue(dst, b.leaves[dfs]), off + 1
-}
-
-func (b *Builder) Encode() []byte {
-	off := 0
-	for off < len(b.nodes) {
-		next := b.measure(off)
-		if next < 0 {
-			break
-		}
-		off = next
-	}
-	off = 0
-	var zv []byte
-	for off < len(b.nodes) {
-		var next int
-		zv, next = b.encode(zv, off)
-		if next < 0 {
-			break
-		}
-		off = next
-	}
-	return zv
-}
-
 // Iter iterates over a sequence of zvals.
 type Iter []byte
 
@@ -175,7 +49,7 @@ func AppendContainer(dst []byte, vals [][]byte) []byte {
 	}
 	var n int
 	for _, v := range vals {
-		n += sizeBytes(v)
+		n += sizeOfByteSlice(v)
 	}
 	dst = AppendUvarint(dst, containerTag(n))
 	for _, v := range vals {
@@ -193,11 +67,11 @@ func AppendValue(dst []byte, val []byte) []byte {
 	return append(dst, val...)
 }
 
-// sizeBytes returns the number of bytes required by AppendValue to represent
+// sizeOfByteSlice returns the number of bytes required by AppendValue to represent
 // the zval in val.
-func sizeBytes(val []byte) int {
+func sizeOfByteSlice(val []byte) int {
 	// This really is correct even when val is nil.
-	return sizeUvarint(valueTag(len(val))) + len(val)
+	return sizeOfUvarint(valueTag(len(val))) + len(val)
 }
 
 // AppendUvarint is like encoding/binary.PutUvarint but appends to dst instead
@@ -210,9 +84,9 @@ func AppendUvarint(dst []byte, u64 uint64) []byte {
 	return append(dst, byte(u64))
 }
 
-// sizeUvarint returns the number of bytes required by AppendUvarint to
+// sizeOfUvarint returns the number of bytes required by AppendUvarint to
 // represent u64.
-func sizeUvarint(u64 uint64) int {
+func sizeOfUvarint(u64 uint64) int {
 	n := 1
 	for u64 >= 0x80 {
 		n++
@@ -228,11 +102,11 @@ func Uvarint(buf []byte) (uint64, int) {
 }
 
 func sizeOfContainer(length int) int {
-	return (sizeUvarint(containerTag(length))) + length
+	return (sizeOfUvarint(containerTag(length))) + length
 }
 
 func sizeOfValue(length int) int {
-	return int(sizeUvarint(valueTag(length))) + length
+	return int(sizeOfUvarint(valueTag(length))) + length
 }
 
 func containerTag(length int) uint64 {
